PiDigit realisation in Python
Made it by creating a **class** based on int. 

# Guide

## Atributes:
- `value`  - storages *int* value of a PiDigit. For example, ```python
x = PiDigit(2,3); x.value``` will be 14
- `pi` - storages special pi of a PiDigit. For example, `PiDigit(2,3)` will store only `.1415927`, while `PiDigit(20,30)` will store `.1415926535897931159979634685441852`
- `startPos` and `endPos` - store start position and end position in pi
## Capabilities
Long story short, PiDigit can do all things that int can in the same way, for example, you can add to PiDigits using `+`. Only thing that differs is that PiDigit `__repr__()` is *PiDigit with startPos - {startPos}, endPos - {endPos} and value equal to {value}*

That's it